# api/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Category:
    properties:
      children:
        description: 关联信息
        items:
          $ref: '#/definitions/domain.Category'
        type: array
      created_at:
        description: 创建时间
        type: string
      id:
        description: 分类ID
        type: string
      inherit_stall:
        description: 是否继承父分类的出品部门ID
        type: boolean
      inherit_tax_rate:
        description: 是否继承父分类的税率ID
        type: boolean
      merchant_id:
        description: 品牌商ID
        type: string
      name:
        description: 分类名称
        type: string
      parent:
        allOf:
        - $ref: '#/definitions/domain.Category'
        description: 父分类
      parent_id:
        description: 父分类ID
        type: string
      product_count:
        description: 关联的商品数量
        type: integer
      sort_order:
        description: 排序，值越小越靠前
        type: integer
      stall:
        allOf:
        - $ref: '#/definitions/domain.Stall'
        description: 出品部门
      stall_id:
        description: 出品部门ID
        type: string
      store_id:
        description: 门店ID
        type: string
      tax_rate:
        allOf:
        - $ref: '#/definitions/domain.TaxFee'
        description: 税率
      tax_rate_id:
        description: 税率ID
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
...
  domain.EffectiveDateType:
    enum:
    - daily
    - custom
    type: string
    x-enum-comments:
      EffectiveDateTypeCustom: 自定义
      EffectiveDateTypeDaily: 按天
    x-enum-varnames:
    - EffectiveDateTypeDaily
    - EffectiveDateTypeCustom
...
  domain.Menu:
    properties:
      created_at:
        description: 创建时间
        type: string
      id:
        description: 菜单ID
        type: string
      item_count:
        description: 菜单项数量
        type: integer
      items:
        description: 菜单项列表
        items:
          $ref: '#/definitions/domain.MenuItem'
        type: array
      merchant_id:
        description: 品牌商ID
        type: string
      name:
        description: 菜单名称
        type: string
      store_count:
        description: 适用门店数量
        type: integer
      store_id:
        description: 门店ID
        type: string
      stores:
        description: 关联信息
        items:
          $ref: '#/definitions/domain.StoreSimple'
        type: array
      updated_at:
        description: 更新时间
        type: string
    type: object
  domain.MenuItem:
    properties:
      base_price:
        description: 基础价（可选，单位：分）
        type: number
      created_at:
        description: 创建时间
        type: string
      id:
        description: 菜单项ID
        type: string
      member_price:
        description: 会员价（可选，单位：分）
        type: number
      menu_id:
        description: 菜单ID
        type: string
      product:
        allOf:
        - $ref: '#/definitions/domain.Product'
        description: 关联信息
      product_id:
        description: 菜品ID
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
  domain.MenuSearchRes:
    properties:
      items:
        items:
          $ref: '#/definitions/domain.Menu'
        type: array
      page:
        description: 页码
        type: integer
      size:
        description: 每页数量
        type: integer
      total:
        description: 总页数
        type: integer
    type: object
...
  domain.Product:
    properties:
      add_sale_quantity:
        description: 加售份数
        type: integer
      attr_relations:
        description: 商品口味做法关联列表
        items:
          $ref: '#/definitions/domain.ProductAttrRelation'
        type: array
      category:
        allOf:
        - $ref: '#/definitions/domain.Category'
        description: 分类
      category_id:
        description: 基础信息
        type: string
      created_at:
        description: 时间戳
        type: string
      delivery_cost_price:
        description: 外卖成本价（可选，单位：分，仅套餐商品使用）
        type: number
      description:
        description: 菜品描述
        type: string
      detail_images:
        description: 详情图片
        items:
          type: string
        type: array
      effective_date_type:
        allOf:
        - $ref: '#/definitions/domain.EffectiveDateType'
        description: 生效日期类型
      effective_end_time:
        description: 生效结束时间
        type: string
      effective_start_time:
        description: 生效开始时间
        type: string
      estimated_cost_price:
        description: 套餐属性（仅套餐商品使用）
        type: number
      groups:
        description: 套餐组列表
        items:
          $ref: '#/definitions/domain.SetMealGroup'
        type: array
      id:
        description: 商品ID
        type: string
      inherit_stall:
        description: 是否继承原出品部门
        type: boolean
      inherit_tax_rate:
        description: 其他信息
        type: boolean
      main_image:
        description: 展示信息
        type: string
      menu_id:
        description: 菜单ID
        type: string
      merchant_id:
        description: 品牌商ID
        type: string
      min_sale_quantity:
        description: 起售份数
        type: integer
      mnemonic:
        description: 助记词
        type: string
      name:
        description: 商品名称
        type: string
      sale_channels:
        description: 售卖渠道
        items:
          $ref: '#/definitions/domain.SaleChannel'
        type: array
      sale_status:
        allOf:
        - $ref: '#/definitions/domain.ProductSaleStatus'
        description: 售卖信息
      shelf_life:
        description: 保质期（单位：天）
        type: integer
      spec_relations:
        description: 关联信息
        items:
          $ref: '#/definitions/domain.ProductSpecRelation'
        type: array
      stall:
        allOf:
        - $ref: '#/definitions/domain.Stall'
        description: 出品部门
      stall_id:
        description: 指定出品部门ID
        type: string
      store_id:
        description: 门店ID
        type: string
      support_types:
        description: 支持类型（堂食、外带）
        items:
          $ref: '#/definitions/domain.ProductSupportType'
        type: array
      tags:
        description: 商品标签列表
        items:
          $ref: '#/definitions/domain.ProductTag'
        type: array
      tax_rate:
        allOf:
        - $ref: '#/definitions/domain.TaxFee'
        description: 税率
      tax_rate_id:
        description: 指定税率ID
        type: string
      type:
        allOf:
        - $ref: '#/definitions/domain.ProductType'
        description: 商品类型
      unit:
        allOf:
        - $ref: '#/definitions/domain.ProductUnit'
        description: 单位
      unit_id:
        description: 属性关联
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
  domain.ProductAttr:
    properties:
      channels:
        description: 售卖渠道列表
        items:
          $ref: '#/definitions/domain.SaleChannel'
        type: array
      created_at:
        description: 创建时间
        type: string
      id:
        description: 口味做法ID
        type: string
      items:
        description: 关联信息
        items:
          $ref: '#/definitions/domain.ProductAttrItem'
        type: array
      merchant_id:
        description: 品牌商ID
        type: string
      name:
        description: 口味做法名称
        type: string
      product_count:
        description: 关联的商品数量（所有项的关联商品数量累加）
        type: integer
      store_id:
        description: 门店ID
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
  domain.ProductAttrItem:
    properties:
      attr_id:
        description: 口味做法ID（外键）
        type: string
      base_price:
        description: 基础加价（单位：分）
        type: number
      created_at:
        description: 创建时间
        type: string
      id:
        description: 口味做法项ID
        type: string
      image:
        description: 图片URL（可选）
        type: string
      name:
        description: 口味做法项名称
        type: string
      product_count:
        description: 关联的商品数量
        type: integer
      updated_at:
        description: 更新时间
        type: string
    type: object
  domain.ProductAttrRelation:
    properties:
      attr:
        allOf:
        - $ref: '#/definitions/domain.ProductAttr'
        description: 关联信息
      attr_id:
        description: 口味做法ID
        type: string
      attr_item:
        allOf:
        - $ref: '#/definitions/domain.ProductAttrItem'
        description: 口味做法项
      attr_item_id:
        description: 口味做法项ID
        type: string
      created_at:
        description: 创建时间
        type: string
      id:
        description: 关联ID
        type: string
      is_default:
        description: 是否默认项
        type: boolean
      product_id:
        description: 商品ID
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
  domain.ProductSaleStatus:
    enum:
    - on_sale
    - off_sale
    type: string
    x-enum-comments:
      ProductSaleStatusOffSale: 停售
      ProductSaleStatusOnSale: 在售
    x-enum-varnames:
    - ProductSaleStatusOnSale
    - ProductSaleStatusOffSale
...
  domain.ProductSpecRelation:
    properties:
      barcode:
        description: 条形码
        type: string
      base_price:
        description: 基础价格（单位：分）
        type: number
      created_at:
        description: 创建时间
        type: string
      estimated_cost_price:
        description: 预估成本价（单位：分，可选）
        type: number
      id:
        description: 关联ID
        type: string
      is_default:
        description: 是否默认项
        type: boolean
      member_price:
        description: 会员价（单位：分，可选）
        type: number
      other_price1:
        description: 其他价格1（单位：分，可选）
        type: number
      other_price2:
        description: 其他价格2（单位：分，可选）
        type: number
      other_price3:
        description: 其他价格3（单位：分，可选）
        type: number
      packing_fee:
        allOf:
        - $ref: '#/definitions/domain.AdditionalFee'
        description: 打包费
      packing_fee_id:
        description: 打包费ID（引用费用配置）
        type: string
      product_id:
        description: 商品ID
        type: string
      spec_id:
        description: 规格ID
        type: string
      spec_name:
        description: 关联信息
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
...
  domain.ProductSupportType:
    enum:
    - dine
    - takeaway
    - delivery
    type: string
    x-enum-comments:
      ProductSupportTypeDelivery: 外卖
      ProductSupportTypeDine: 堂食
      ProductSupportTypeTakeaway: 外带
    x-enum-varnames:
    - ProductSupportTypeDine
    - ProductSupportTypeTakeaway
    - ProductSupportTypeDelivery
  domain.ProductTag:
    properties:
      created_at:
        description: 创建时间
        type: string
      id:
        description: 标签ID
        type: string
      merchant_id:
        description: 品牌商ID
        type: string
      name:
        description: 标签名称
        type: string
      product_count:
        description: 关联的商品数量
        type: integer
      store_id:
        description: 门店ID
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
...
  domain.ProductType:
    enum:
    - normal
    - set_meal
    type: string
    x-enum-comments:
      ProductTypeNormal: 普通商品
      ProductTypeSetMeal: 套餐商品
    x-enum-varnames:
    - ProductTypeNormal
    - ProductTypeSetMeal
  domain.ProductUnit:
    properties:
      created_at:
        description: 创建时间
        type: string
      id:
        description: 单位ID
        type: string
      merchant_id:
        description: 品牌商ID
        type: string
      name:
        description: 单位名称
        type: string
      product_count:
        description: 关联的商品数量
        type: integer
      store_id:
        description: 门店ID
        type: string
      type:
        allOf:
        - $ref: '#/definitions/domain.ProductUnitType'
        description: 单位类型：quantity（数量单位）、weight（重量单位）
      updated_at:
        description: 更新时间
        type: string
    type: object
...
  domain.RouterMenu:
    properties:
      component:
        description: 前端组件路径
        type: string
      created_at:
        type: string
      enabled:
        description: 是否启用
        type: boolean
      icon:
        description: 菜单图标
        type: string
      id:
        type: string
      layer:
        description: 菜单层级
        type: integer
      name:
        description: 菜单名称
        type: string
      parent_id:
        description: 父级菜单ID，根菜单为 uuid.Nil
        type: string
      path:
        description: 菜单路径
        type: string
      sort:
        description: 菜单排序
        type: integer
      updated_at:
        type: string
      user_type:
        allOf:
        - $ref: '#/definitions/domain.UserType'
        description: 类型
    type: object
  domain.SaleChannel:
    enum:
    - POS
    - Mobile
    - Scan
    - SelfService
    - ThirdParty
    - POS
    - Mobile
    - Scan
    - SelfService
    - ThirdParty
    type: string
    x-enum-comments:
      PaymentMethodDisplayChannelMobileOrdering: 移动点餐
      PaymentMethodDisplayChannelPOS: POS
      PaymentMethodDisplayChannelScanOrdering: 扫码点餐
      PaymentMethodDisplayChannelSelfService: 自助点餐
      PaymentMethodDisplayChannelThirdPartyDelivery: 三方外卖
      SaleChannelMobileOrdering: 移动点餐
      SaleChannelPOS: POS
      SaleChannelScanOrdering: 扫码点餐
      SaleChannelSelfService: 自助点餐
      SaleChannelThirdPartyDelivery: 三方外卖
    x-enum-varnames:
    - PaymentMethodDisplayChannelPOS
    - PaymentMethodDisplayChannelMobileOrdering
    - PaymentMethodDisplayChannelScanOrdering
    - PaymentMethodDisplayChannelSelfService
    - PaymentMethodDisplayChannelThirdPartyDelivery
    - SaleChannelPOS
    - SaleChannelMobileOrdering
    - SaleChannelScanOrdering
    - SaleChannelSelfService
    - SaleChannelThirdPartyDelivery
  domain.SetMealDetail:
    properties:
      created_at:
        description: 创建时间
        type: string
      group_id:
        description: 套餐组ID（外键）
        type: string
      id:
        description: 详情ID
        type: string
      is_default:
        description: 是否默认（必选，每个套餐组中只能有一个默认项）
        type: boolean
      optional_product_ids:
        description: 备选商品ID列表（可选，多选）
        items:
          type: string
        type: array
      optional_products:
        description: 备选商品
        items:
          $ref: '#/definitions/domain.Product'
        type: array
      product:
        allOf:
        - $ref: '#/definitions/domain.Product'
        description: 关联信息
      product_id:
        description: 商品ID（外键，引用普通商品）
        type: string
      quantity:
        description: 数量（必选，必须为正整数）
        type: integer
      updated_at:
        description: 更新时间
        type: string
    type: object
  domain.SetMealGroup:
    properties:
      created_at:
        description: 创建时间
        type: string
      details:
        description: 关联信息
        items:
          $ref: '#/definitions/domain.SetMealDetail'
        type: array
      id:
        description: 套餐组ID
        type: string
      name:
        description: 套餐组名称
        type: string
      product_id:
        description: 套餐商品ID（外键）
        type: string
      selection_type:
        allOf:
        - $ref: '#/definitions/domain.SetMealGroupSelectionType'
        description: 点单限制：fixed（固定分组）、optional（可选套餐）
      updated_at:
        description: 更新时间
        type: string
    type: object
  domain.SetMealGroupSelectionType:
    enum:
    - fixed
    - optional
    type: string
    x-enum-comments:
      SetMealGroupSelectionTypeFixed: 固定分组
      SetMealGroupSelectionTypeOptional: 可选套餐
    x-enum-varnames:
    - SetMealGroupSelectionTypeFixed
    - SetMealGroupSelectionTypeOptional
...
  domain.Stall:
    properties:
      created_at:
        description: 创建时间
        type: string
      enabled:
        description: 使用状态
        type: boolean
      id:
        type: string
      merchant_id:
        description: 商户 ID
        type: string
      name:
        description: 出品部门名称
        type: string
      print_type:
        allOf:
        - $ref: '#/definitions/domain.StallPrintType'
        description: 打印类型 ：receipt/label
      sort_order:
        description: 排序
        type: integer
      stall_type:
        allOf:
        - $ref: '#/definitions/domain.StallType'
        description: system/brand/store
      store_id:
        description: 门店 ID
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
...
  domain.StoreSimple:
    properties:
      id:
        type: string
      store_name:
        description: 门店名称
        type: string
    type: object
...
  domain.TaxFee:
    properties:
      created_at:
        type: string
      default_tax:
        description: 是否默认税费
        type: boolean
      id:
        type: string
      merchant_id:
        description: 品牌商 ID
        type: string
      name:
        description: 税费名称
        type: string
      store_id:
        description: 门店 ID
        type: string
      tax_code:
        description: 税费代码
        type: string
      tax_fee_type:
        allOf:
        - $ref: '#/definitions/domain.TaxFeeType'
        description: 税费类型：商户/门店
      tax_rate:
        description: 税率，6% -> 0.06
        type: number
      tax_rate_type:
        allOf:
        - $ref: '#/definitions/domain.TaxRateType'
        description: 税率类型：统一比例/自定义比例
      updated_at:
        type: string
    type: object
...
  domain.UserType:
    enum:
    - admin
    - backend
    - store
    - admin
    - backend
    - store
    type: string
    x-enum-comments:
      UserTypeAdmin: admin表用户
      UserTypeBackend: backend用户
      UserTypeStore: store用户
    x-enum-varnames:
    - RoleTypeAdmin
    - RoleTypeBackend
    - RoleTypeStore
...
  types.MenuCreateReq:
    properties:
      items:
        description: 菜品列表（必选，至少一个）
        items:
          $ref: '#/definitions/types.MenuItemReq'
        minItems: 1
        type: array
      name:
        description: 菜单名称（必选）
        maxLength: 255
        type: string
      store_ids:
        description: 门店ID列表（必选，多选）
        items:
          type: string
        minItems: 1
        type: array
    required:
    - items
    - name
    - store_ids
    type: object
  types.MenuItemReq:
    properties:
      base_price:
        description: 基础价（可选，单位：分）
        type: number
      member_price:
        description: 会员价（可选，单位：分）
        type: number
      product_id:
        description: 菜品ID（必选）
        type: string
    required:
    - product_id
    type: object
  types.MenuUpdateReq:
    properties:
      items:
        description: 菜品列表（必选，至少一个）
        items:
          $ref: '#/definitions/types.MenuItemReq'
        minItems: 1
        type: array
      name:
        description: 菜单名称（必选）
        maxLength: 255
        type: string
      store_ids:
        description: 门店ID列表（必选，多选）
        items:
          type: string
        minItems: 1
        type: array
    required:
    - items
    - name
    - store_ids
    type: object
...
  types.RoleMenusResp:
    properties:
      paths:
        items:
          type: string
        type: array
    type: object
...
  types.RouterMenuCreateReq:
    properties:
      component:
        description: 组件路径
        maxLength: 255
        type: string
      enabled:
        description: 是否启用
        type: boolean
      icon:
        description: 菜单图标
        maxLength: 500
        type: string
      name:
        description: 菜单名称
        maxLength: 100
        type: string
      parent_id:
        description: 父级菜单ID
        type: string
      path:
        description: 路由路径
        maxLength: 255
        type: string
      sort:
        description: 排序
        type: integer
    required:
    - name
    type: object
  types.RouterMenuListResp:
    properties:
      menus:
        items:
          $ref: '#/definitions/domain.RouterMenu'
        type: array
      total:
        type: integer
    type: object
  types.RouterMenuUpdateReq:
    properties:
      component:
        description: 组件路径
        maxLength: 255
        type: string
      enabled:
        description: 是否启用
        type: boolean
      icon:
        description: 菜单图标
        maxLength: 500
        type: string
      name:
        description: 菜单名称
        maxLength: 100
        type: string
      parent_id:
        description: 父级菜单ID
        type: string
      path:
        description: 路由路径
        maxLength: 255
        type: string
      sort:
        description: 排序
        type: integer
    required:
    - name
    type: object
...
  types.SetMenusReq:
    properties:
      paths:
        items:
          type: string
        type: array
    required:
    - paths
    type: object
...
info:
  contact: {}
  description: 供品牌商后台调用.
  title: 品牌商后台 API
  version: "1.0"
paths:
...
  /common/role/{id}/menus:
    get:
      consumes:
      - application/json
      description: 分页或非分页获取指定角色的菜单路径列表
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.RoleMenusResp'
      security:
      - BearerAuth: []
      summary: 角色菜单列表
      tags:
      - 角色管理
    post:
      consumes:
      - application/json
      description: 为指定角色设置菜单路径（交集保留，新增/删除按 paths 调整）
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: string
      - description: 设置菜单请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.SetMenusReq'
      produces:
      - application/json
      responses:
        "200":
          description: No Content
      security:
      - BearerAuth: []
      summary: 设置角色菜单
      tags:
      - 角色管理
  /common/router_menu:
    get:
      description: 分页查询菜单
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.RouterMenuListResp'
      security:
      - BearerAuth: []
      summary: 菜单列表
      tags:
      - 菜单管理
    post:
      consumes:
      - application/json
      description: 新建一个菜单
      parameters:
      - description: 创建菜单请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.RouterMenuCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: No Content
      security:
      - BearerAuth: []
      summary: 创建菜单
      tags:
      - 菜单管理
  /common/router_menu/{id}:
    delete:
      description: 删除指定菜单
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: No Content
      security:
      - BearerAuth: []
      summary: 删除菜单
      tags:
      - 菜单管理
    get:
      description: 查询指定菜单
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.RouterMenu'
      security:
      - BearerAuth: []
      summary: 获取菜单详情
      tags:
      - 菜单管理
    put:
      consumes:
      - application/json
      description: 修改指定菜单
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新菜单请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.RouterMenuUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: No Content
      security:
      - BearerAuth: []
      summary: 更新菜单
      tags:
      - 菜单管理
  /common/router_menu/{id}/disable:
    put:
      description: 禁用指定菜单
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: No Content
      security:
      - BearerAuth: []
      summary: 禁用菜单
      tags:
      - 菜单管理
...
  /menu:
    get:
      parameters:
      - description: 菜单名称（模糊匹配）
        in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/domain.MenuSearchRes'
      security:
      - BearerAuth: []
      summary: 查询菜单列表
      tags:
      - 菜单管理
    post:
      parameters:
      - description: 请求信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.MenuCreateReq'
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: 创建菜单
      tags:
      - 菜单管理
  /menu/{id}:
    delete:
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: 删除菜单
      tags:
      - 菜单管理
    get:
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/domain.Menu'
      security:
      - BearerAuth: []
      summary: 获取菜单详情
      tags:
      - 菜单管理
    put:
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: string
      - description: 请求信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.MenuUpdateReq'
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: 更新菜单
      tags:
      - 菜单管理
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"